+++ crates/nostr-relay-pool/src/prelude.rs
@@ -9,6 +9,7 @@
 #![doc(hidden)]
 
 // External crates
+pub use async_utility::futures_util::StreamExt; // Needed for `RelayPool::stream_events_of`
 pub use nostr::prelude::*;
 pub use nostr_database::*;